fix: add shared validated config, fix client using server functionality
