+++ .env.sample
@@ -1,3 +1,3 @@
 DATA_STORE=json
-NEXT_PUBLIC_STRAPI_API_URL=http://localhost:1337/api
+STRAPI_API_URL=http://localhost:1337/api
 STRAPI_API_TOKEN=token
