+++ components/organisms/ThemeToggle.tsx
@@ -2,7 +2,7 @@
 
 import { Moon, Sun } from 'lucide-react'
 import { useTheme } from 'next-themes'
-import React from 'react'
+import { useEffect, useState } from 'react'
 
 import { Button } from '@/components/atoms/Button'
 import {
@@ -13,7 +13,16 @@ import {
 } from '@/components/molecules/DropdownMenu'
 
 export const ThemeToggle = () => {
+    const [mounted, setMounted] = useState(false)
     const { setTheme } = useTheme()
+
+    useEffect(() => {
+        setMounted(true)
+    }, [])
+
+    if (!mounted) {
+        return null
+    }
 
     return (
         <DropdownMenu>
