+++ lib/clients/index.ts
@@ -1,6 +1,6 @@
+import { DATA_STORE_PROVIDERS, config } from '@/env'
 import { createJSONClient } from '@/lib/clients/json/jsonClient'
 import { createStrapiClient } from '@/lib/clients/strapi/strapiClient'
-import { DATA_STORE_PROVIDERS } from '@/next.config' // TODO: fix config, share among server and client
 
 import type { CMSClient } from '@/lib/clients/types'
 
@@ -21,6 +21,4 @@ const createCMSClient = (cms: DataStoreProvider): CMSClient => {
     return clientCreator()
 }
 
-export const client = createCMSClient(
-    (process.env.DATA_STORE || 'json') as 'strapi' | 'json' // TODO: fix config, share among server and client
-)
+export const client = createCMSClient(config.DATA_STORE || 'json')
