+++ lib/serverUtils.ts
@@ -1,0 +1,12 @@
+import { config } from '@/env'
+
+export const getImageUrl = (url: string) => {
+    if (!url) return ''
+    try {
+        const { STRAPI_API_URL: baseUrl } = config
+        return new URL(url, baseUrl).toString()
+    } catch (error) {
+        console.error('Failed to construct image URL:', error)
+        return ''
+    }
+}
