+++ lib/utils.ts
@@ -6,17 +6,6 @@ export const classNames = (...inputs: ClassValue[]) => {
     return twMerge(clsx(inputs))
 }
 
-export const getImageUrl = (url: string) => {
-    if (!url) return ''
-    try {
-        const baseUrl = process.env.NEXT_PUBLIC_STRAPI_API_URL
-        return new URL(url, baseUrl).toString()
-    } catch (error) {
-        console.error('Failed to construct image URL:', error)
-        return ''
-    }
-}
-
 const DATE_FORMAT = 'DD MMM, YYYY'
 
 export const formatDate = (
