+++ package.json
@@ -52,6 +52,7 @@
         "eslint": "9.17.0",
         "eslint-config-next": "15.1.3",
         "eslint-plugin-import": "^2.31.0",
+        "eslint-plugin-react-hooks": "^5.1.0",
         "eslint-plugin-unicorn": "^56.0.1",
         "husky": "^9.1.7",
         "json-schema-to-ts": "^3.1.1",
